MAGECLOUD-5490: Capture MySQL statistics
