+++ etc/crontab.xml
@@ -6,8 +6,8 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Cron:etc/crontab.xsd">
-    <group id="cloud_analytics">
-        <job name="cloud_components_log_process_list" instance="Magento\CloudComponents\Model\Cron\LogProcessList" method="execute">
+    <group id="default">
+        <job name="cloud_components_db_health" instance="Magento\CloudComponents\Model\Cron\DbHealthList" method="execute">
             <schedule>* * * * *</schedule>
         </job>
     </group>
